Move lib files to utils folder
